Fix warning for 32bit builds
When building for 32bit platforms (e.g. Android armv7a), then printing
qsizetype value using printf's %lld format generates a warning because
qsizetype is just an int. This breaks the build, and can't be fixed
by using another placeholder without producing a warning on 64bit
builds. So fix it by casting the qsizetype always to 64bit signed int.

+++ src/widgets/graphicsview/qgraphicsanchorlayout_p.cpp
@@ -529,7 +529,7 @@ void AnchorData::dump(int indent) {
         p->secondEdge->dump(indent+2);
     } else if (type == Sequential) {
         const auto *s = static_cast<SequentialAnchorData *>(this);
-        qDebug("%*s type: sequential(%lld):", indent, "", s->m_edges.size());
+        qDebug("%*s type: sequential(%lld):", indent, "", qint64(s->m_edges.size()));
         for (AnchorData *e : s->m_edges)
             e->dump(indent + 2);
     } else {
